fixed swappable tasks by replacing draggable and droppable with sortable
